requirements.txt
# Use python==3.8.20 
# Core ML libraries

scikit-learn==1.3.2
xgboost==2.1.4
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1
joblib==1.4.2

# Visualization
matplotlib==3.7.5
seaborn==0.13.2
#altair==5.4.1
plotly==5.24.1


#flask==3.0.3
flask==2.2.5

# MLflow for experiment tracking
mlflow==2.17.2
mlflow-skinny==2.17.2

# Data drift & monitoring
#evidently==0.4.15

# Streamlit dashboard
#streamlit==1.40.1

# Optional utilities
#pydantic==1.10.13
#requests==2.32.4
#python-dotenv==1.0.1
# Postgres Ingestion Using sqlalchemy
psycopg2-binary==2.9.9
SQLAlchemy==1.4.50
shap